Organized code + data files and added comments to make the code more legible.
