+++ data/evaluation tests/random.french
@@ -1,0 +1,2 @@
+j'ai été licencié .	i was fired .	0-0 1-1 2-2 3-3	
+personne n'est chez lui .	no one is home .	0-1 1-0 2-3 3-3 4-4
