--- data/random.french
@@ -1,2 +1,0 @@
-j'ai été licencié .	i was fired .	0-0 1-1 2-2 3-3	0-0 1-0 2-0 3-0
-personne n'est chez lui .	no one is home .	0-1 1-0 2-3 3-3 4-4	0-0 1-0 2-0 3-0 4-0
