--- data/test.alignments
@@ -1,1 +1,0 @@
-	0-0 1-0 2-3 3-3 4-3 5-4	0-0 1-0 2-0 3-0 4-0 5-0
