--- data/test_sentences.txt
@@ -1,4 +1,0 @@
-Wow!	Ça alors !
-Grab that.	Saisis-toi de ça !
-He's lazy.	Il est paresseux.
-Forget it!	Oublie !
